# lib/api-zod/src/generated/types/workTreeNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkTreeNode {
  id: number;
  runId: number;
  parentId: number | null;
  title: string;
  detail: string;
  kind: string;
  status: string;
  depth: number;
  position: number;
  result: string;
  verification: string;
  attempts: number;
  trace: string;
  role: string;
  createdAt: string;
  updatedAt: string;
}